% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\docType{data}
\name{assets}
\alias{assets}
\title{Assets Data}
\format{
A vector with 10 numeric values representing assets figures.
}
\source{
text book
}
\usage{
assets
}
\description{
This data set contains data of assets in billions
}
\keyword{datasets}
