% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T1_5.R
\docType{data}
\name{T1_5}
\alias{T1_5}
\title{T1-5 data}
\format{
A data frame with 42 rows and 7 columns
\describe{
  \item{V1}{Wind}
  \item{V2}{Solar radiation}
  \item{V3}{CO}
  \item{V4}{NO}
  \item{V5}{NO2}
  \item{V6}{O3}
  \item{V7}{HC}
  ...
}
}
\source{
text book
}
\usage{
T1_5
}
\description{
This is a data about air-pollution. This contains 42 measurements with 7 variables
}
\keyword{datasets}
